backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "diffusers>=0.35.2",
    "dotenv>=0.9.9",
    "fastapi>=0.121.3",
    "huggingface-hub>=0.36.0",
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "numpy>=2.3.5",
    "onnxruntime-gpu>=1.23.2",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-multipart>=0.0.20",
    "realesrgan>=0.3.0",
    "rembg>=2.0.68",
    "safetensors>=0.7.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]
